Implement LoansTable summary
