+++ src/pages/OffersPage/components/ActiveTabContent/components/LoansTable/LoansTable.module.less
@@ -71,3 +71,69 @@
 .liquidated {
   background: var(--additional-red-secondary) !important;
 }
+
+//? Summary styles
+
+.summaryContainer {
+  display: flex;
+  align-items: center;
+  padding: 12px 16px;
+  border-top: 1px solid var(--bg-border);
+  gap: 80px;
+  width: 100%;
+
+  @media (max-width: 640px) {
+    gap: 8px;
+  }
+}
+
+.infoRow {
+  display: flex;
+  align-items: center;
+  gap: 16px;
+  white-space: nowrap;
+
+  &:last-child {
+    margin-left: auto;
+  }
+
+  @media (max-width: 960px) {
+    flex-direction: column;
+    width: 100%;
+  }
+}
+
+.loansContainer {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
+
+.loansValueText {
+  font: var(--important-text-xll);
+
+  @media (max-width: 960px) {
+    display: none;
+  }
+}
+
+.loansInfoContainer {
+  @media (max-width: 960px) {
+    & > div {
+      text-align: center;
+    }
+  }
+}
+
+.summaryButton {
+  width: 128px;
+
+  @media (max-width: 960px) {
+    width: 100%;
+  }
+}
+
+.terminateButton:not(:disabled) {
+  background: var(--additional-lava-secondary) !important;
+  border-color: var(--additional-lava-primary) !important;
+}
