+++ src/pages/OffersPage/components/OffersTabContent/components/Summary/Summary.module.less
@@ -64,7 +64,7 @@
     display: none;
   }
 }
-.loansInfoContainer,
+
 .loansInterestContainer {
   @media (max-width: 960px) {
     & > div {
@@ -79,9 +79,4 @@
   @media (max-width: 960px) {
     width: 100%;
   }
-}
-
-.terminateButton:not(:disabled) {
-  background: var(--additional-lava-secondary) !important;
-  border-color: var(--additional-lava-primary) !important;
 }
